Refactor: Simplify package imports and Game class access
This commit addresses the following:

1. Configured `pyproject.toml` to recognize the `src` directory as the root for package discovery. This allows you to import the main package as `bracket_city_mcp` instead of `src.bracket_city_mcp` when the package is installed.
   - Added `[tool.setuptools.packages.find]` with `where = ["src"]`.

2. Simplified the import of the `Game` class.
   - Added `from .game import Game` to `src/bracket_city_mcp/game/__init__.py`.
   - This allows `from bracket_city_mcp.game import Game` instead of `from bracket_city_mcp.game.game import Game`.

3. Updated import statements in example and test files to use the new, cleaner paths.
   - `examples/play_game.py` now uses `from bracket_city_mcp.game import Game` and has the `sys.path` manipulation for `SRC_DIR` removed.
   - `tests/test_game.py` now uses `from bracket_city_mcp.game import Game`.

All existing tests pass with these changes, and the example script runs correctly, confirming the desired import behavior has been achieved while retaining the `src` directory structure.
