+++ pyproject.toml
@@ -21,3 +21,6 @@ dev = [
 
 [tool.pytest.ini_options]
 asyncio_mode = "auto"
+
+[tool.setuptools.packages.find]
+where = ["src"]
