+++ src/bracket_city_mcp/game/__init__.py
@@ -1,0 +1,1 @@
+from .game import Game
